Added BOM header check for snap_from_csv
Without the BOM header check, this causes issues with CSV files that were made in a spreadsheet software and have BOM at the start. Causing the parser to read the first cell wrong. While it seems like nothing out of the ordinary, as it renders just fine. It adds an invisible character. Breaking all sorts of things internally.

This fix implements a BOM header check, advances the seek by 3 and modifies the size/word_start variables to compensate.